pyproject.toml
[project]
name = "fairmaps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartogram>=1.0.0",
    "dotenv>=0.9.9",
    "geopandas>=1.1.1",
    "matplotlib>=3.10.7",
    "pygris>=0.2.0",
    "pytidycensus>=1.2.3",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ipython>=9.8.0",
    "pytest>=9.0.1",
]
